+++ app/lib/theme/widgets/custom/CommunitySourceWidget.tsx
@@ -43,6 +43,9 @@ const CommunitySourceWidget: React.FC<CommunitySourceWidgetProps> = (props) => {
     value: geographicNameId,
     label: bcGeographicName,
   });
+  const [economicRegionInputValue, setEconomicRegionInputValue] = useState('');
+  const [regionalDistrictInputValue, setRegionalDistrictValue] = useState('');
+  const [geographicNameInputValue, setGeographicNameInputValue] = useState('');
 
   useEffect(() => {
     setSelectedEconomicRegion(economicRegion);
@@ -119,9 +122,12 @@ const CommunitySourceWidget: React.FC<CommunitySourceWidgetProps> = (props) => {
             setSelectedEconomicRegion(val);
           }
         }}
+        onInputChange={(e, val) => {
+          setEconomicRegionInputValue(val);
+        }}
         style={{ width: '200px' }}
         value={selectedEconomicRegion}
-        inputValue={selectedEconomicRegion ?? ''}
+        inputValue={economicRegionInputValue}
         options={economicRegionOptions}
         getOptionLabel={(option) => option}
         renderInput={(params) => (
@@ -147,8 +153,11 @@ const CommunitySourceWidget: React.FC<CommunitySourceWidgetProps> = (props) => {
             setSelectedRegionalDistrict(val);
           }
         }}
+        onInputChange={(e, val) => {
+          setRegionalDistrictValue(val);
+        }}
         value={selectedRegionalDistrict}
-        inputValue={selectedRegionalDistrict ?? ''}
+        inputValue={regionalDistrictInputValue}
         options={
           regionalDistrictOptions[selectedEconomicRegion]
             ? [...regionalDistrictOptions[selectedEconomicRegion]]
@@ -177,6 +186,11 @@ const CommunitySourceWidget: React.FC<CommunitySourceWidgetProps> = (props) => {
             label="Geographic Name"
           />
         )}
+        renderOption={(renderProps, option) => (
+          <li {...renderProps} key={option.value}>
+            {option.label}
+          </li>
+        )}
         options={getGeographicNameOptions(
           selectedRegionalDistrict,
           selectedEconomicRegion
@@ -189,7 +203,7 @@ const CommunitySourceWidget: React.FC<CommunitySourceWidgetProps> = (props) => {
           return option.label ?? '';
         }}
         value={selectedGeographicName}
-        inputValue={selectedGeographicName?.label ?? ''}
+        inputValue={geographicNameInputValue}
         onChange={(e, val, reason) => {
           if (reason === 'clear') {
             setSelectedGeographicName({ value: null, label: '' });
@@ -204,6 +218,9 @@ const CommunitySourceWidget: React.FC<CommunitySourceWidgetProps> = (props) => {
               geographicNameId: val.value,
             });
           }
+        }}
+        onInputChange={(e, val) => {
+          setGeographicNameInputValue(val);
         }}
       />
       {!rowId && (